requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pandas
openpyxl
scikit-learn
joblib
surprise
